requirements.txt
discord.py
asyncpg
redis[hiredis]
aiohttp
python-dotenv
quart
validators